+++ plugins/receiver/grpc/envoymetricsv2/receiver_test.go
@@ -1,0 +1,67 @@
+// Licensed to Apache Software Foundation (ASF) under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Apache Software Foundation (ASF) licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package envoymetricsv2
+
+import (
+	"context"
+	"strconv"
+	"testing"
+
+	"skywalking.apache.org/repo/goapi/proto/envoy/api/v2/core"
+
+	v2 "skywalking.apache.org/repo/goapi/proto/envoy/service/metrics/v2"
+	v1 "skywalking.apache.org/repo/goapi/satellite/data/v1"
+
+	"google.golang.org/grpc"
+
+	_ "github.com/apache/skywalking-satellite/internal/satellite/test"
+	receiver_grpc "github.com/apache/skywalking-satellite/plugins/receiver/grpc"
+)
+
+func TestReceiver_RegisterHandler(t *testing.T) {
+	recConf := make(map[string]string, 2)
+	recConf["limit_count"] = "1"
+	recConf["flush_time"] = "1000"
+	receiver_grpc.TestReceiverWithConfig(new(Receiver), recConf, func(t *testing.T, sequence int, conn *grpc.ClientConn, ctx context.Context) string {
+		client := v2.NewMetricsServiceClient(conn)
+		data := initData(sequence)
+		collect, err := client.StreamMetrics(ctx)
+		if err != nil {
+			t.Fatalf("cannot open the stream send mode: %v", err)
+		}
+		if err := collect.Send(data); err != nil {
+			t.Fatalf("cannot send the data to the server: %v", err)
+		}
+		if err := collect.CloseSend(); err != nil {
+			t.Fatalf("cannot close the stream mode: %v", err)
+		}
+		return data.String()
+	}, func(data *v1.SniffData) string {
+		return data.GetEnvoyMetricsV2List().Messages[0].String()
+	}, t)
+}
+
+func initData(sequence int) *v2.StreamMetricsMessage {
+	return &v2.StreamMetricsMessage{
+		Identifier: &v2.StreamMetricsMessage_Identifier{
+			Node: &core.Node{
+				Id: "test" + strconv.Itoa(sequence),
+			},
+		},
+	}
+}