+++ CHANGES.md
@@ -8,6 +8,7 @@ Release Notes.
 * Support load-balance GRPC client with the static server list.
 * Support load-balance GRPC client with the Kubernetes selector.
 * Support transmit Envoy ALS v2/v3 protocol.
+* Support transmit Envoy Metrics v2/v3 protocol.
 
 #### Bug Fixes
 * Fix errors when converting meter data from histogram and summary.[#75](https://github.com/apache/skywalking-satellite/pull/75)